requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-cloud-aiplatform[preview]>=1.49.0
google-cloud-storage>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
ffmpeg-python>=0.2.0
opencv-python>=4.9.0
pydub>=0.25.1
scikit-image>=0.23.0
numpy>=1.26.0

